Update fork from base

+++ hi_dsp/modules/Modulators.h
@@ -1001,7 +1001,7 @@ private:
 			auto bMod = index - bIndex;
 			auto before = data[bIndex];
 
-			data[bIndex] = before & ~(uint64)((1 << bMod));
+			data[bIndex] = before & ~(uint64)((uint64)(1) << bMod);
 
 			if(before != data[bIndex])
 				numBitsSet = jmax(0, numBitsSet - 1);
